requirements_fixed.txt
# Core ML packages with compatible versions
openai>=1.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.0
langchain>=0.1.0
# Fixed sentence transformers dependencies
numpy<2.0.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.21.0
huggingface-hub>=0.19.0
# Additional dependencies
pandas>=1.5.0
scikit-learn>=1.2.0
jinja2>=3.1.0
requests>=2.28.0
